APAP/requirements.txt
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.2
django-allauth==0.39.1
django-appconf==1.0.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-image-cropping==1.2.0
django-openid-auth==0.15
django-simple-email-confirmation==0.60
easy-thumbnails==2.6
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
oauthlib==3.0.1
openpyxl==2.6.2
Pillow==6.0.0
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
xlrd==1.2.0
